Replace FLAMO direct path with numeric dss_to_pr_direct
