+++ src/pyFDN/__init__.py
@@ -69,10 +69,10 @@ __all__ = [
     "dss_to_flamo",
     "dss_to_impz",
     "dss_to_pr",
+    "dss_to_pr_direct",
     "flamo_to_pr",
     "flamo_extract_pr_decomposition",
     "dss_to_pr_flamo",
-    "dss_to_pr_flamo_direct",
     "dss_to_ss",
     "dss_to_tf",
     "impz_to_res",
@@ -256,12 +256,12 @@ from .translate.dss_to_flamo import dss_to_flamo
 from .translate.dss_to_ss import dss_to_ss
 from .translate.dss_to_impz import dss_to_impz
 from .translate.dss_to_pr import dss_to_pr
+from .translate.dss_to_pr_direct import dss_to_pr_direct
 from .translate.dss_to_pr_flamo import (
     dss_to_pr_flamo,
     flamo_extract_pr_decomposition,
     flamo_to_pr,
 )
-from .translate.dss_to_pr_flamo_direct import dss_to_pr_flamo_direct
 from .translate.dss_to_tf import dss_to_tf
 from .translate.impz_to_res import impz_to_res
 from .translate.mtf_to_impz import mtf_to_impz
