+++ tests/test_dss_to_pr_direct.py
@@ -1,4 +1,4 @@
-"""Tests for direct FLAMO-only dss_to_pr_flamo_direct path."""
+"""Tests for DSS-only direct dss_to_pr_direct path."""
 
 from __future__ import annotations
 
@@ -6,32 +6,19 @@ import numpy as np
 
 from pyFDN.generate.random_orthogonal import random_orthogonal
 from pyFDN.translate.dss_to_impz import dss_to_impz
-from pyFDN.translate.dss_to_pr_flamo_direct import dss_to_pr_flamo_direct
+from pyFDN.translate.dss_to_pr_direct import dss_to_pr_direct
 from pyFDN.translate.pr_to_impz import pr_to_impz
 
 
-class ConstantFeedbackGraph:
-    def __init__(self, matrix: np.ndarray):
-        self._matrix = np.asarray(matrix, dtype=np.float64)
-        self.input_channels = self._matrix.shape[1]
-        self.output_channels = self._matrix.shape[0]
-
-    def at(self, z):
-        return self._matrix
-
-    def der(self, z):
-        return np.zeros_like(self._matrix)
-
-
-def test_dss_to_pr_flamo_direct_reconstructs_ir_eig_mode():
+def test_dss_to_pr_direct_reconstructs_ir_eig_mode():
     delays = np.array([4, 5, 6, 7], dtype=int)
     a = 0.55 * random_orthogonal(delays.size)
     b = np.eye(delays.size, 1)
     c = np.eye(1, delays.size)
     d = np.zeros((1, 1))
 
-    residues, poles, direct, is_pair, _ = dss_to_pr_flamo_direct(
-        delays, ConstantFeedbackGraph(a), b, c, d, mode="eig"
+    residues, poles, direct, is_pair, _ = dss_to_pr_direct(
+        delays, a, b, c, d, mode="eig"
     )
 
     ir_len = 512
@@ -41,15 +28,15 @@ def test_dss_to_pr_flamo_direct_reconstructs_ir_eig_mode():
     np.testing.assert_allclose(ir_modal, ir_time, rtol=1e-7, atol=1e-8)
 
 
-def test_dss_to_pr_flamo_direct_roots_mode_runs():
+def test_dss_to_pr_direct_roots_mode_runs():
     delays = np.array([4, 5, 6], dtype=int)
     a = 0.5 * random_orthogonal(delays.size)
     b = np.eye(delays.size, 1)
     c = np.eye(1, delays.size)
     d = np.zeros((1, 1))
 
-    residues, poles, direct, is_pair, meta = dss_to_pr_flamo_direct(
-        delays, ConstantFeedbackGraph(a), b, c, d, mode="roots"
+    residues, poles, direct, is_pair, meta = dss_to_pr_direct(
+        delays, a, b, c, d, mode="roots"
     )
 
     assert residues.ndim == 3
